feat!: compute width/height internally
Instead of requiring the user to specifically set width and height,
we compute it from the <svg> element's size. The user can control this
through CSS on the <svg> element (via className, or styled components).
The Foundation is only rendered when width and height are available.

There is still an optional initialWidth/initialHeight which can be
used for initial rendering.

BREAKING:
 - `onReady` no longer passes `visibleArea`, but an object with
   data to be able to compute any transformation (more generic).
   Check the docs for `onReady` for an example how to compute the
   visible area.
 - `width` and `height` properties removed from `Foundation`,
   you can still set the size for the <svg> element via CSS
   (if auto size does not work for you).